Enhancements (#128)
* in-progress, Grab map provider

* in-progress, Grab map provider

* handled undefined credentials case when stack is switched

* updated tests

* updated constants

* in progress, grab map and style support

* fixed ALS-1408

* fixed info content which was overlapping with logo

* fix for ALS-1252

* fix for ALS-1408

* in-progress, updates to Grab feature

* in-progress

* [ALS-1311] added tests and finalised functionality

* updated readme

* fixed build issue

* resolved bug and updated CF template link in README

* updated e2e test

* updated e2e test

* removed @tanstack/react-query

* removed @tanstack/react-query

* added tooltips and resolved innerHtml issue in console

* resoolved location when disconnecting stack while using Grab map provider

* bug fixes

* bug fixes

* attribution text color handling for different styles, removed bias position when searching with Grab map provider

* removed commented styles

* attribution fixes

* padding added to cater long suggestions in search

* fixed zoom issue for popup when clicking on search suggestion and renamed Grab to GrabMaps

* added await for E2E test

* added feedback from QA

* removed outline for info icon on Popup

* feedback from QA

* removed log

* enabled Route Options for walking when using Grab, disabled locate me icon UI fix

* style updates

* added CF templates

* synced upstream

* merged feature_ALS-1311

* simplified DemoPage component

* simplified DemoPage component

* updated default CF template

* removed unecessary prop

---------
