+++ src/atomicui/pages/DemoPage/DemoPage.tsx
@@ -29,50 +29,33 @@ import {
 	UnauthSimulation
 } from "@demo/atomicui/organisms";
 import { DemoPlaceholderPage } from "@demo/atomicui/pages";
-import { showToast } from "@demo/core";
-import { appConfig, regionsData } from "@demo/core/constants";
+import { appConfig } from "@demo/core/constants";
 import BottomSheetHeights from "@demo/core/constants/bottomSheetHeights";
 import {
 	useAmplifyAuth,
 	useAmplifyMap,
-	useAws,
 	useAwsGeofence,
-	useAwsIot,
 	useAwsPlace,
 	useAwsRoute,
 	useAwsTracker,
+	useCredsManager,
+	useMapManager,
 	usePersistedData,
 	useRecordViewPage
 } from "@demo/hooks";
 import useBottomSheet from "@demo/hooks/useBottomSheet";
 import useDeviceMediaQuery from "@demo/hooks/useDeviceMediaQuery";
-import {
-	EsriMapEnum,
-	GrabMapEnum,
-	HereMapEnum,
-	MapProviderEnum,
-	MapStyleFilterTypes,
-	MenuItemEnum,
-	ShowStateType,
-	ToastType
-} from "@demo/types";
-import { EventTypeEnum, OpenDataMapEnum, ResponsiveUIEnum, TriggeredByEnum } from "@demo/types/Enums";
-import { record } from "@demo/utils/analyticsUtils";
+import { GrabMapEnum, MapProviderEnum, MapStyleFilterTypes, MenuItemEnum, ShowStateType } from "@demo/types";
+import { ResponsiveUIEnum, TriggeredByEnum } from "@demo/types/Enums";
 import { errorHandler } from "@demo/utils/errorHandler";
-import { getCurrentLocation } from "@demo/utils/getCurrentLocation";
 import { Signer } from "aws-amplify";
-import { differenceInMilliseconds } from "date-fns";
 import { LngLatBoundsLike } from "mapbox-gl";
-import { omit } from "ramda";
 import { useTranslation } from "react-i18next";
 import {
 	AttributionControl,
 	GeolocateControl,
 	GeolocateControlRef,
-	GeolocateErrorEvent,
-	GeolocateResultEvent,
 	Map,
-	MapLayerMouseEvent,
 	MapRef,
 	NavigationControl
 } from "react-map-gl";
@@ -81,15 +64,10 @@ import "./styles.scss";
 import { RefHandles } from "react-spring-bottom-sheet/dist/types";
 
 const {
-	POOLS,
-	PERSIST_STORAGE_KEYS: { SHOULD_CLEAR_CREDENTIALS, GEO_LOCATION_ALLOWED, FASTEST_REGION },
-	ROUTES: { DEMO },
-	MAP_RESOURCES: { MAX_BOUNDS, AMAZON_HQ, GRAB_SUPPORTED_AWS_REGIONS },
-	LINKS: { AMAZON_LOCATION_TERMS_AND_CONDITIONS, AWS_LOCATION },
-	GET_PARAMS: { DATA_PROVIDER }
+	MAP_RESOURCES: { MAX_BOUNDS },
+	LINKS: { AMAZON_LOCATION_TERMS_AND_CONDITIONS, AWS_LOCATION }
 } = appConfig;
 const initShow = {
-	gridLoader: true,
 	sidebar: false,
 	routeBox: false,
 	signInModal: false,
@@ -100,9 +78,6 @@ const initShow = {
 	stylesCard: false,
 	authTrackerDisclaimerModal: false,
 	about: false,
-	grabDisclaimerModal: false,
-	openDataDisclaimerModal: false,
-	mapStyle: undefined,
 	unauthGeofenceBox: false,
 	unauthTrackerBox: false,
 	unauthSimulationBounds: false,
@@ -111,10 +86,6 @@ const initShow = {
 	startUnauthSimulation: false,
 	openFeedbackModal: false
 };
-let interval: NodeJS.Timer | undefined;
-let timeout: NodeJS.Timer | undefined;
-const searchParams = new URLSearchParams(window.location.search);
-let switchToMapProvider = searchParams.get(DATA_PROVIDER);
 const peggedRemValue = 13;
 const extraGeoLocateTop = 2.6;
 
@@ -124,8 +95,6 @@ const DemoPage: React.FC = () => {
 	const [isUnauthNotifications, setUnauthIsNotifications] = React.useState(false);
 	const [confirmCloseUnauthSimulation, setConfirmCloseUnauthSimulation] = React.useState(false);
 	const [searchValue, setSearchValue] = React.useState("");
-	const [doNotAskGrabDisclaimer, setDoNotAskGrabDisclaimer] = React.useState(false);
-	const [doNotAskOpenDataDisclaimer, setDoNotAskOpenDataDisclaimer] = React.useState(false);
 	const [triggerOnClose, setTriggerOnClose] = React.useState(false);
 	const [triggerOnReset, setTriggerOnReset] = React.useState(false);
 	const [expandRouteOptionsMobile, setExpandRouteOptionsMobile] = React.useState(false);
@@ -139,77 +108,64 @@ const DemoPage: React.FC = () => {
 	const [searchBoxValue, setSearchBoxValue] = React.useState("");
 	const mapViewRef = useRef<MapRef | null>(null);
 	const geolocateControlRef = useRef<GeolocateControlRef | null>(null);
-	const {
-		credentials,
-		getCurrentUserCredentials,
-		clearCredentials,
-		region,
-		authTokens,
-		setAuthTokens,
-		onLogout,
-		handleCurrentSession,
-		switchToGrabMapRegionStack,
-		isUserAwsAccountConnected,
-		switchToDefaultRegionStack,
-		handleStackRegion
-	} = useAmplifyAuth();
-	const { locationClient, createLocationClient, iotClient, createIotClient, resetStore: resetAwsStore } = useAws();
-	const { attachPolicy } = useAwsIot();
+	const { credentials, region, isUserAwsAccountConnected } = useAmplifyAuth();
 	const {
 		mapProvider: currentMapProvider,
 		mapStyle: currentMapStyle,
 		currentLocationData,
 		setMapProvider,
-		setMapStyle,
 		isCurrentLocationDisabled,
-		setIsCurrentLocationDisabled,
-		setCurrentLocation,
-		viewpoint,
-		setViewpoint,
-		autoMapUnit,
-		setAutomaticMapUnit
+		viewpoint
 	} = useAmplifyMap();
-	const { setMarker, marker, selectedMarker, suggestions, bound, clearPoiList, zoom, setZoom, setSelectedMarker } =
-		useAwsPlace();
-	const { routeData, directions, resetStore: resetAwsRouteStore, setRouteData } = useAwsRoute();
+	const { selectedMarker, suggestions, bound, clearPoiList, zoom, setZoom } = useAwsPlace();
+	const { routeData, directions, resetStore: resetAwsRouteStore } = useAwsRoute();
 	const { resetStore: resetAwsGeofenceStore } = useAwsGeofence();
-	const { isEditingRoute, trackerPoints, setTrackerPoints, resetStore: resetAwsTrackingStore } = useAwsTracker();
-	const {
-		showWelcomeModal,
-		setShowWelcomeModal,
-		doNotAskGrabDisclaimerModal,
-		setDoNotAskGrabDisclaimerModal,
-		doNotAskOpenDataDisclaimerModal,
-		setDoNotAskOpenDataDisclaimerModal,
-		setSettingsOptions
-	} = usePersistedData();
+	const { isEditingRoute, resetStore: resetAwsTrackingStore } = useAwsTracker();
+	const { showWelcomeModal, setShowWelcomeModal, setSettingsOptions } = usePersistedData();
 	const { isDesktop, isMobile, isTablet } = useDeviceMediaQuery();
+	const { setUI, ui, bottomSheetCurrentHeight = 0, setBottomSheetHeight, setBottomSheetMinHeight } = useBottomSheet();
+	const { clearCredsAndLocationClient } = useCredsManager();
 	const {
-		setUI,
-		ui,
-		bottomSheetCurrentHeight = 0,
-		setBottomSheetHeight,
-		setBottomSheetMinHeight,
-		setBottomSheetOpen
-	} = useBottomSheet();
+		gridLoader,
+		setGridLoader,
+		tempMapStyle,
+		setTempMapStyle,
+		grabDisclaimerModal,
+		setGrabDisclaimerModal,
+		// doNotAskGrabDisclaimer,
+		setDoNotAskGrabDisclaimer,
+		openDataDisclaimerModal,
+		setOpenDataDisclaimerModal,
+		// doNotAskOpenDataDisclaimer,
+		setDoNotAskOpenDataDisclaimer,
+		isGrabVisible,
+		onLoad,
+		getCurrentGeoLocation,
+		onGeoLocate,
+		onGeoLocateError,
+		handleMapClick,
+		handleCurrentLocationAndViewpoint,
+		handleOpenDataMapChange,
+		resetAppState,
+		handleGrabMapChange,
+		onMapProviderChange,
+		onMapStyleChange
+	} = useMapManager({
+		mapViewRef,
+		geolocateControlRef,
+		isUnauthGeofenceBoxOpen: show.unauthGeofenceBox,
+		isUnauthTrackerBoxOpen: show.unauthTrackerBox,
+		isAuthGeofenceBoxOpen: show.authGeofenceBox,
+		isSettingsOpen: show.settings,
+		isRouteBoxOpen: show.routeBox,
+		closeRouteBox: () => setShow(s => ({ ...s, routeBox: false })),
+		resetAppStateCb: () => setShow(s => ({ ...initShow, stylesCard: s.stylesCard, settings: s.settings })),
+		setUnauthSimulationExitModal: (b: boolean) => setShow(s => ({ ...s, unauthSimulationExitModal: b }))
+	});
 	const { t, i18n } = useTranslation();
 	const langDir = i18n.dir();
 	const isLtr = langDir === "ltr";
-	const shouldClearCredentials = localStorage.getItem(SHOULD_CLEAR_CREDENTIALS) === "true";
 	const geoLocateTopValue = `-${bottomSheetCurrentHeight / peggedRemValue + extraGeoLocateTop}rem`;
-	const fallbackRegion = Object.values(POOLS)[0];
-	const fastestRegion = localStorage.getItem(FASTEST_REGION) ?? fallbackRegion;
-	const defaultRegion = regionsData.find(option => option.value === fastestRegion) as { value: string; label: string };
-
-	const isGrabAvailableInRegion = useMemo(() => !!region && GRAB_SUPPORTED_AWS_REGIONS.includes(region), [region]);
-
-	const isGrabVisible = useMemo(
-		() =>
-			!show.unauthGeofenceBox &&
-			!show.unauthTrackerBox &&
-			(!isUserAwsAccountConnected || (isUserAwsAccountConnected && isGrabAvailableInRegion)),
-		[show.unauthGeofenceBox, show.unauthTrackerBox, isUserAwsAccountConnected, isGrabAvailableInRegion]
-	);
 
 	useEffect(() => {
 		let previousWidth = document.body.clientWidth;
@@ -233,10 +189,6 @@ const DemoPage: React.FC = () => {
 		};
 	}, []);
 
-	useEffect(() => {
-		autoMapUnit.selected && setAutomaticMapUnit();
-	}, [autoMapUnit.selected, setAutomaticMapUnit]);
-
 	const handleResetCallback = useCallback(
 		function handleReset() {
 			setSearchValue("");
@@ -244,117 +196,6 @@ const DemoPage: React.FC = () => {
 		},
 		[setSearchValue, setSelectedFilters]
 	);
-
-	const clearCredsAndLocationClient = useCallback(() => {
-		clearCredentials();
-		resetAwsStore();
-	}, [clearCredentials, resetAwsStore]);
-
-	if (shouldClearCredentials || (!!credentials && !credentials?.identityId)) {
-		localStorage.removeItem(SHOULD_CLEAR_CREDENTIALS);
-		clearCredsAndLocationClient();
-	}
-
-	/* Fetch the current user credentials */
-	useEffect(() => {
-		if (credentials?.identityId && credentials?.expiration) {
-			const now = new Date();
-			const expiration = new Date(credentials.expiration);
-
-			if (now > expiration) {
-				/* If the credentials are expired, clear them and the location client */
-				clearCredsAndLocationClient();
-			} else {
-				/* If the credentials are not expired, set the refresh interval/timeout */
-				interval && clearInterval(interval);
-				timeout && clearTimeout(timeout);
-
-				if (credentials.authenticated) {
-					/* If the credentials are authenticated, set the refresh interval */
-					interval = setInterval(() => {
-						handleCurrentSession(resetAwsStore);
-					}, 10 * 60 * 1000);
-				} else {
-					/* If the credentials are not authenticated, set the refresh timeout */
-					timeout = setTimeout(() => {
-						clearCredsAndLocationClient();
-					}, differenceInMilliseconds(new Date(credentials.expiration || 0), new Date()));
-				}
-			}
-		} else {
-			/* If the credentials are not present, fetch them */
-			getCurrentUserCredentials();
-		}
-	}, [credentials, getCurrentUserCredentials, handleCurrentSession, resetAwsStore, clearCredsAndLocationClient]);
-
-	/* Instantiate location and iot client from aws-sdk whenever the credentials change */
-	useEffect(() => {
-		if (credentials && region) {
-			!locationClient && createLocationClient(credentials, region);
-			!iotClient && createIotClient(credentials, region);
-		}
-	}, [credentials, locationClient, createLocationClient, region, iotClient, createIotClient]);
-
-	const _onLogout = useCallback(async () => {
-		await onLogout();
-		clearCredentials();
-		resetAwsStore();
-	}, [onLogout, clearCredentials, resetAwsStore]);
-
-	/* Fired when user logs in or logs out */
-	useEffect(() => {
-		const searchParams = new URLSearchParams(window.location.search);
-		const code = searchParams.get("code");
-		const state = searchParams.get("state");
-		const sign_out = searchParams.get("sign_out");
-
-		/* After login */
-		if (code && state && !authTokens) {
-			window.history.replaceState(undefined, "", DEMO);
-			setAuthTokens({ code, state });
-			record(
-				[{ EventType: EventTypeEnum.SIGN_IN_SUCCESSFUL, Attributes: {} }],
-				["userAWSAccountConnectionStatus", "userAuthenticationStatus"]
-			);
-			setTimeout(() => clearCredsAndLocationClient(), 0);
-		}
-
-		/* After logout */
-		if (sign_out === "true") {
-			window.history.replaceState(undefined, "", DEMO);
-			!!credentials?.authenticated && !authTokens && _onLogout();
-		}
-	}, [setAuthTokens, clearCredsAndLocationClient, credentials, authTokens, _onLogout]);
-
-	const _attachPolicy = useCallback(async () => {
-		if (credentials?.identityId && credentials?.expiration) {
-			const now = new Date();
-			const expiration = new Date(credentials.expiration);
-
-			if (now > expiration) {
-				/* If the credentials are expired, clear them and the location client */
-				clearCredsAndLocationClient();
-			} else {
-				if (!!credentials.authenticated && !!authTokens) {
-					await attachPolicy(credentials.identityId);
-				} else if (!isUserAwsAccountConnected && currentMapProvider !== MapProviderEnum.GRAB) {
-					await attachPolicy(credentials.identityId, true);
-				}
-			}
-		}
-	}, [
-		credentials,
-		clearCredsAndLocationClient,
-		authTokens,
-		attachPolicy,
-		isUserAwsAccountConnected,
-		currentMapProvider
-	]);
-
-	/* Attach IoT policy to authenticated user to ensure successful websocket connection */
-	useEffect(() => {
-		_attachPolicy();
-	}, [_attachPolicy]);
 
 	useEffect(() => {
 		if (selectedMarker) {
@@ -415,116 +256,6 @@ const DemoPage: React.FC = () => {
 		if (directions) setShow(s => ({ ...s, routeBox: true }));
 	}, [directions, show]);
 
-	const onLoad = useCallback(() => {
-		clearPoiList();
-		!isCurrentLocationDisabled && geolocateControlRef.current?.trigger();
-		routeData && setRouteData(routeData);
-	}, [clearPoiList, isCurrentLocationDisabled, routeData, setRouteData]);
-
-	const getCurrentGeoLocation = useCallback(() => {
-		if (!region) {
-			return;
-		}
-
-		if (isGrabAvailableInRegion) {
-			if (isCurrentLocationDisabled) {
-				showToast({
-					content: t("show_toast__grab_not_supported.text"),
-					type: ToastType.INFO
-				});
-				mapViewRef.current?.flyTo({ center: [AMAZON_HQ.SG.longitude, AMAZON_HQ.SG.latitude], zoom: 15 });
-			} else {
-				getCurrentLocation(setCurrentLocation, setViewpoint, currentMapProvider, setIsCurrentLocationDisabled);
-			}
-		} else {
-			getCurrentLocation(setCurrentLocation, setViewpoint, currentMapProvider, setIsCurrentLocationDisabled);
-		}
-	}, [
-		region,
-		isGrabAvailableInRegion,
-		isCurrentLocationDisabled,
-		t,
-		setCurrentLocation,
-		setViewpoint,
-		currentMapProvider,
-		setIsCurrentLocationDisabled
-	]);
-
-	useEffect(() => {
-		if ("permissions" in navigator && region) {
-			navigator.permissions.query({ name: "geolocation" }).then(({ state }) => {
-				const permissionAllowed = localStorage.getItem(GEO_LOCATION_ALLOWED) || "no";
-
-				if (permissionAllowed === "no" && state === "granted") {
-					localStorage.setItem(GEO_LOCATION_ALLOWED, "yes");
-					getCurrentGeoLocation();
-				}
-			});
-		}
-	}, [getCurrentGeoLocation, region]);
-
-	const onGeoLocate = useCallback(
-		({ coords: { latitude, longitude } }: GeolocateResultEvent) => {
-			if (routeData) {
-				resetAwsRouteStore();
-				setShow(s => ({ ...s, routeBox: false }));
-
-				setTimeout(() => {
-					setViewpoint({ latitude, longitude });
-					setCurrentLocation({ currentLocation: { latitude, longitude }, error: undefined });
-				}, 0);
-			} else {
-				setViewpoint({ latitude, longitude });
-				setCurrentLocation({ currentLocation: { latitude, longitude }, error: undefined });
-			}
-		},
-		[resetAwsRouteStore, routeData, setCurrentLocation, setViewpoint]
-	);
-
-	const onGeoLocateError = useCallback(
-		(e: GeolocateErrorEvent) => {
-			setCurrentLocation({ currentLocation: undefined, error: { ...omit(["type", "target"], e) } });
-
-			if (e.code === e.PERMISSION_DENIED) {
-				localStorage.setItem(GEO_LOCATION_ALLOWED, "no");
-				showToast({
-					content: t("show_toast__lpd.text"),
-					type: ToastType.ERROR
-				});
-			} else if (e.code === e.POSITION_UNAVAILABLE) {
-				showToast({
-					content: t("show_toast__lpu.text"),
-					type: ToastType.ERROR
-				});
-			}
-		},
-		[setCurrentLocation, t]
-	);
-
-	const handleMapClick = ({ lngLat }: MapLayerMouseEvent) => {
-		if (lngLat && !show.unauthGeofenceBox && !show.unauthTrackerBox) {
-			const { lat: latitude, lng: longitude } = lngLat;
-
-			if (!show.routeBox && !show.authGeofenceBox && !show.settings && !isEditingRoute) {
-				marker && setMarker(undefined);
-				selectedMarker && setSelectedMarker(undefined);
-				setTimeout(() => setMarker({ latitude, longitude }), 0);
-			}
-
-			if (isEditingRoute) {
-				if (trackerPoints) {
-					trackerPoints.length < 25
-						? setTrackerPoints([longitude, latitude])
-						: showToast({ content: t("show_toast__route_waypoint_restriction.text"), type: ToastType.WARNING });
-				} else {
-					setTrackerPoints([longitude, latitude]);
-				}
-			}
-
-			mapViewRef?.current?.flyTo({ center: lngLat });
-		}
-	};
-
 	const onEnableTracking = () => {
 		clearPoiList();
 		resetAwsRouteStore();
@@ -562,348 +293,6 @@ const DemoPage: React.FC = () => {
 			return { url: newUrl };
 		},
 		[region, credentials]
-	);
-
-	const resetAppState = useCallback(() => {
-		clearPoiList();
-		resetAwsRouteStore();
-		resetAwsGeofenceStore();
-		resetAwsTrackingStore();
-		setShow(s => ({ ...initShow, stylesCard: s.stylesCard, settings: s.settings }));
-	}, [clearPoiList, resetAwsRouteStore, resetAwsGeofenceStore, resetAwsTrackingStore]);
-
-	const handleCurrentLocationAndViewpoint = useCallback(
-		(switchToGrab = true) => {
-			if (switchToGrab) {
-				/* When switching to Grab */
-				if (currentLocationData?.currentLocation) {
-					/* If current location data exists */
-					const { latitude, longitude } = currentLocationData.currentLocation;
-					const [westBound, southBound, eastBound, northBound] = MAX_BOUNDS.GRAB;
-					const isWithinGrabBounds =
-						latitude >= southBound && latitude <= northBound && longitude >= westBound && longitude <= eastBound;
-
-					if (!isWithinGrabBounds) {
-						/* If current location lies outside Grab MAX_BOUNDS */
-						setIsCurrentLocationDisabled(true);
-						setViewpoint({ latitude: AMAZON_HQ.SG.latitude, longitude: AMAZON_HQ.SG.longitude });
-						setZoom(15);
-						mapViewRef.current?.flyTo({
-							center: [AMAZON_HQ.SG.longitude, AMAZON_HQ.SG.latitude]
-						});
-					}
-				} else {
-					setViewpoint({ latitude: AMAZON_HQ.SG.latitude, longitude: AMAZON_HQ.SG.longitude });
-					setZoom(15);
-					mapViewRef.current?.flyTo({
-						center: [AMAZON_HQ.SG.longitude, AMAZON_HQ.SG.latitude]
-					});
-				}
-			} else {
-				/* When switching between other map providers or styles */
-				if (currentLocationData?.currentLocation) {
-					const { latitude, longitude } = currentLocationData.currentLocation;
-					isCurrentLocationDisabled && setIsCurrentLocationDisabled(false);
-					setViewpoint({ latitude, longitude });
-					setZoom(15);
-					mapViewRef.current?.flyTo({ center: [longitude, latitude] });
-					setTimeout(() => {
-						geolocateControlRef.current?.trigger();
-					}, 3000);
-				} else {
-					setViewpoint({ latitude: AMAZON_HQ.US.latitude, longitude: AMAZON_HQ.US.longitude });
-					setZoom(15);
-					mapViewRef.current?.flyTo({
-						center: [AMAZON_HQ.US.longitude, AMAZON_HQ.US.latitude]
-					});
-				}
-			}
-		},
-		[currentLocationData, setViewpoint, setZoom, setIsCurrentLocationDisabled, isCurrentLocationDisabled]
-	);
-
-	const handleOpenDataMapChange = useCallback(
-		(mapStyle?: OpenDataMapEnum) => {
-			if (doNotAskOpenDataDisclaimerModal) setDoNotAskOpenDataDisclaimerModal(!doNotAskOpenDataDisclaimer);
-
-			setShow(s => ({ ...s, openDataDisclaimerModal: false, gridLoader: true }));
-
-			if (currentMapProvider === MapProviderEnum.GRAB && !isUserAwsAccountConnected && fastestRegion !== region) {
-				switchToDefaultRegionStack();
-				resetAwsStore();
-				setIsCurrentLocationDisabled(false);
-				!isDesktop && setBottomSheetOpen();
-			}
-
-			setMapProvider(MapProviderEnum.OPEN_DATA);
-			setMapStyle(
-				(typeof mapStyle === "string" ? mapStyle : undefined) ||
-					(show.mapStyle ? show.mapStyle : OpenDataMapEnum.OPEN_DATA_STANDARD_LIGHT)
-			);
-			handleCurrentLocationAndViewpoint(false);
-		},
-		[
-			doNotAskOpenDataDisclaimerModal,
-			setDoNotAskOpenDataDisclaimerModal,
-			doNotAskOpenDataDisclaimer,
-			currentMapProvider,
-			isUserAwsAccountConnected,
-			fastestRegion,
-			region,
-			setMapProvider,
-			setMapStyle,
-			show.mapStyle,
-			handleCurrentLocationAndViewpoint,
-			switchToDefaultRegionStack,
-			resetAwsStore,
-			setIsCurrentLocationDisabled,
-			isDesktop,
-			setBottomSheetOpen
-		]
-	);
-
-	const handleGrabMapChange = useCallback(
-		(mapStyle?: GrabMapEnum) => {
-			if (doNotAskGrabDisclaimerModal) setDoNotAskGrabDisclaimerModal(!doNotAskGrabDisclaimer);
-			else setShow(s => ({ ...s, grabDisclaimerModal: false }));
-
-			if (!isUserAwsAccountConnected && !isGrabAvailableInRegion) {
-				switchToGrabMapRegionStack();
-				resetAwsStore();
-				!isDesktop && setBottomSheetOpen();
-			}
-
-			setMapProvider(MapProviderEnum.GRAB);
-			setMapStyle(
-				(typeof mapStyle === "string" ? mapStyle : undefined) ||
-					(show.mapStyle ? show.mapStyle : GrabMapEnum.GRAB_STANDARD_LIGHT)
-			);
-			setShow(s => ({ ...s, gridLoader: true }));
-			resetAppState();
-			handleCurrentLocationAndViewpoint();
-		},
-		[
-			doNotAskGrabDisclaimerModal,
-			setDoNotAskGrabDisclaimerModal,
-			doNotAskGrabDisclaimer,
-			isUserAwsAccountConnected,
-			isGrabAvailableInRegion,
-			setMapProvider,
-			setMapStyle,
-			show.mapStyle,
-			resetAppState,
-			handleCurrentLocationAndViewpoint,
-			switchToGrabMapRegionStack,
-			resetAwsStore,
-			isDesktop,
-			setBottomSheetOpen
-		]
-	);
-
-	const onMapProviderChange = useCallback(
-		(mapProvider: MapProviderEnum, triggeredBy: TriggeredByEnum) => {
-			setShow(s => ({ ...s, gridLoader: true }));
-
-			if (mapProvider === MapProviderEnum.OPEN_DATA) {
-				if (doNotAskOpenDataDisclaimerModal) {
-					/* Switching from different map provider and style to OpenData map provider and style */
-					setShow(s => ({ ...s, openDataDisclaimerModal: true }));
-				} else handleOpenDataMapChange();
-			} else if (mapProvider === MapProviderEnum.GRAB) {
-				if (doNotAskGrabDisclaimerModal) {
-					/* Switching from different map provider and style to Grab map provider and style */
-					setShow(s => ({ ...s, grabDisclaimerModal: true }));
-				} else handleGrabMapChange();
-			} else {
-				if (currentMapProvider === MapProviderEnum.GRAB) {
-					/* Switching from Grab map provider to different map provider and style */
-					if (!isUserAwsAccountConnected && fastestRegion !== region) {
-						switchToDefaultRegionStack();
-						resetAwsStore();
-						setIsCurrentLocationDisabled(false);
-
-						if (!isDesktop) {
-							setTimeout(() => {
-								setBottomSheetMinHeight(window.innerHeight * 0.4 - 10);
-								setBottomSheetHeight(window.innerHeight * 0.4);
-							}, 1000);
-
-							setTimeout(() => {
-								setBottomSheetMinHeight(BottomSheetHeights.explore.min);
-								setBottomSheetHeight(window.innerHeight);
-							}, 1200);
-						}
-					}
-
-					setMapProvider(mapProvider);
-					setMapStyle(mapProvider === MapProviderEnum.ESRI ? EsriMapEnum.ESRI_LIGHT : HereMapEnum.HERE_EXPLORE);
-					handleCurrentLocationAndViewpoint(false);
-				} else {
-					/* Switching between Esri and HERE map provider and style */
-					setMapProvider(mapProvider);
-					setMapStyle(mapProvider === MapProviderEnum.ESRI ? EsriMapEnum.ESRI_LIGHT : HereMapEnum.HERE_EXPLORE);
-				}
-
-				resetAppState();
-			}
-
-			record([
-				{ EventType: EventTypeEnum.MAP_PROVIDER_CHANGE, Attributes: { provider: String(mapProvider), triggeredBy } }
-			]);
-		},
-		[
-			doNotAskOpenDataDisclaimerModal,
-			handleOpenDataMapChange,
-			doNotAskGrabDisclaimerModal,
-			handleGrabMapChange,
-			currentMapProvider,
-			resetAppState,
-			isUserAwsAccountConnected,
-			fastestRegion,
-			region,
-			setMapProvider,
-			setMapStyle,
-			handleCurrentLocationAndViewpoint,
-			switchToDefaultRegionStack,
-			resetAwsStore,
-			setIsCurrentLocationDisabled,
-			isDesktop,
-			setBottomSheetMinHeight,
-			setBottomSheetHeight
-		]
-	);
-
-	/* Handle search query params for map provider */
-	useEffect(() => {
-		if (switchToMapProvider) {
-			const { ESRI, HERE, GRAB, OPEN_DATA } = MapProviderEnum;
-
-			if (![ESRI, HERE, GRAB, "GrabMaps", OPEN_DATA].includes(switchToMapProvider)) {
-				/* Invalid search query param for map provider */
-				onMapProviderChange(currentMapProvider, TriggeredByEnum.DEMO_PAGE);
-			} else if (currentMapProvider !== switchToMapProvider) {
-				/* If search query param exist, update map provider based on search query param */
-				if (["Grab", "GrabMaps"].includes(switchToMapProvider)) {
-					isGrabVisible ? onMapProviderChange(GRAB, TriggeredByEnum.DEMO_PAGE) : setMapProvider(currentMapProvider);
-				} else {
-					onMapProviderChange(switchToMapProvider as MapProviderEnum, TriggeredByEnum.DEMO_PAGE);
-				}
-			}
-			switchToMapProvider = null;
-		} else if (!location.search.includes(`${DATA_PROVIDER}=`)) {
-			/* If search query param doesn't exist, update search query param based on current map provider */
-			setMapProvider(currentMapProvider);
-		}
-	}, [currentMapProvider, isGrabVisible, setMapProvider, onMapProviderChange]);
-
-	/* Handled stack region and cloudformation link */
-	useEffect(() => {
-		if (defaultRegion) {
-			currentMapProvider === MapProviderEnum.GRAB
-				? handleStackRegion({
-						value: "ap-southeast-1",
-						label: "regions__ap_southeast_1.text"
-				  })
-				: handleStackRegion(defaultRegion);
-		}
-	}, [defaultRegion, currentMapProvider, handleStackRegion]);
-
-	const onMapStyleChange = useCallback(
-		(mapStyle: EsriMapEnum | HereMapEnum | GrabMapEnum | OpenDataMapEnum) => {
-			const splitArr = mapStyle.split(".");
-			const mapProviderFromStyle = splitArr[splitArr.length - 2] as MapProviderEnum;
-			setShow(s => ({ ...s, gridLoader: true }));
-
-			if (
-				(currentMapProvider === MapProviderEnum.ESRI && mapProviderFromStyle === MapProviderEnum.ESRI) ||
-				(currentMapProvider === MapProviderEnum.HERE && mapProviderFromStyle === MapProviderEnum.HERE) ||
-				(currentMapProvider === MapProviderEnum.GRAB && mapProviderFromStyle === MapProviderEnum.GRAB) ||
-				(currentMapProvider === MapProviderEnum.OPEN_DATA && mapProviderFromStyle === MapProviderEnum.OPEN_DATA)
-			) {
-				/* No map provider switch required */
-				setMapStyle(mapStyle);
-			} else if (mapProviderFromStyle === MapProviderEnum.OPEN_DATA) {
-				/* Switching to OpenData map provider from different map provider and style */
-				if (doNotAskOpenDataDisclaimerModal) {
-					setTimeout(
-						() => setShow(s => ({ ...s, openDataDisclaimerModal: true, mapStyle: mapStyle as OpenDataMapEnum })),
-						0
-					);
-				} else {
-					handleOpenDataMapChange(mapStyle as OpenDataMapEnum);
-				}
-			} else {
-				if (currentMapProvider === MapProviderEnum.GRAB) {
-					/* Switching from Grab map provider to different map provider and style */
-					if (!isUserAwsAccountConnected && fastestRegion !== region) {
-						switchToDefaultRegionStack();
-						resetAwsStore();
-						setIsCurrentLocationDisabled(false);
-
-						if (!isDesktop) {
-							setTimeout(() => {
-								setBottomSheetMinHeight(window.innerHeight * 0.4 - 10);
-								setBottomSheetHeight(window.innerHeight * 0.4);
-							}, 1000);
-
-							setTimeout(() => {
-								setBottomSheetMinHeight(BottomSheetHeights.explore.min);
-								setBottomSheetHeight(window.innerHeight);
-							}, 1200);
-						}
-					}
-
-					setMapProvider(mapProviderFromStyle);
-					setMapStyle(mapStyle);
-					handleCurrentLocationAndViewpoint(false);
-				} else if (mapProviderFromStyle === MapProviderEnum.GRAB) {
-					/* Switching from different map provider and style to Grab map provider and style */
-					if (doNotAskGrabDisclaimerModal) {
-						setTimeout(
-							() =>
-								setShow(s => ({
-									...s,
-									grabDisclaimerModal: true,
-									mapStyle: mapStyle as GrabMapEnum
-								})),
-							0
-						);
-					} else {
-						show.unauthGeofenceBox || show.unauthTrackerBox
-							? setShow(s => ({ ...s, unauthSimulationExitModal: true, mapStyle: mapStyle as GrabMapEnum }))
-							: handleGrabMapChange(mapStyle as GrabMapEnum);
-					}
-				} else {
-					/* Switching between Esri and HERE map provider and style */
-					setMapProvider(mapProviderFromStyle);
-					setMapStyle(mapStyle);
-				}
-
-				!show.unauthGeofenceBox && !show.unauthTrackerBox && resetAppState();
-			}
-		},
-		[
-			currentMapProvider,
-			setMapStyle,
-			doNotAskOpenDataDisclaimerModal,
-			handleOpenDataMapChange,
-			show.unauthGeofenceBox,
-			show.unauthTrackerBox,
-			resetAppState,
-			isUserAwsAccountConnected,
-			fastestRegion,
-			region,
-			setMapProvider,
-			handleCurrentLocationAndViewpoint,
-			switchToDefaultRegionStack,
-			resetAwsStore,
-			setIsCurrentLocationDisabled,
-			isDesktop,
-			setBottomSheetMinHeight,
-			setBottomSheetHeight,
-			doNotAskGrabDisclaimerModal,
-			handleGrabMapChange
-		]
 	);
 
 	const searchBoxEl = useCallback(
@@ -1052,11 +441,11 @@ const DemoPage: React.FC = () => {
 				onLoad={onLoad}
 				onZoom={({ viewState }) => setZoom(viewState.zoom)}
 				onError={error => errorHandler(error.error)}
-				onIdle={() => show.gridLoader && setShow(s => ({ ...s, gridLoader: false }))}
+				onIdle={() => gridLoader && setGridLoader(false)}
 				transformRequest={transformRequest}
 				attributionControl={false}
 			>
-				<View className={show.gridLoader ? "loader-container" : ""}>
+				<View className={gridLoader ? "loader-container" : ""}>
 					{isDesktop && (
 						<>
 							{show.sidebar && (
@@ -1114,9 +503,9 @@ const DemoPage: React.FC = () => {
 									onCloseSidebar={() => setShow(s => ({ ...s, sidebar: false }))}
 									onOpenSignInModal={() => setShow(s => ({ ...s, signInModal: true }))}
 									isGrabVisible={isGrabVisible}
-									showGrabDisclaimerModal={show.grabDisclaimerModal}
-									showOpenDataDisclaimerModal={show.openDataDisclaimerModal}
-									onShowGridLoader={() => setShow(s => ({ ...s, gridLoader: true }))}
+									showGrabDisclaimerModal={grabDisclaimerModal}
+									showOpenDataDisclaimerModal={openDataDisclaimerModal}
+									onShowGridLoader={() => setGridLoader(true)}
 									handleMapStyleChange={onMapStyleChange}
 									searchValue={searchValue}
 									setSearchValue={setSearchValue}
@@ -1221,9 +610,9 @@ const DemoPage: React.FC = () => {
 						onCloseSidebar={() => setShow(s => ({ ...s, sidebar: false }))}
 						onOpenSignInModal={() => setShow(s => ({ ...s, signInModal: true }))}
 						isGrabVisible={isGrabVisible}
-						showGrabDisclaimerModal={show.grabDisclaimerModal}
-						showOpenDataDisclaimerModal={show.openDataDisclaimerModal}
-						onShowGridLoader={() => setShow(s => ({ ...s, gridLoader: true }))}
+						showGrabDisclaimerModal={grabDisclaimerModal}
+						showOpenDataDisclaimerModal={openDataDisclaimerModal}
+						onShowGridLoader={() => setGridLoader(true)}
 						handleMapStyleChange={onMapStyleChange}
 						searchValue={searchValue}
 						setSearchValue={setSearchValue}
@@ -1308,9 +697,9 @@ const DemoPage: React.FC = () => {
 						onCloseSidebar={() => setShow(s => ({ ...s, sidebar: false }))}
 						onOpenSignInModal={() => setShow(s => ({ ...s, signInModal: true }))}
 						isGrabVisible={isGrabVisible}
-						showGrabDisclaimerModal={show.grabDisclaimerModal}
-						showOpenDataDisclaimerModal={show.openDataDisclaimerModal}
-						onShowGridLoader={() => setShow(s => ({ ...s, gridLoader: true }))}
+						showGrabDisclaimerModal={grabDisclaimerModal}
+						showOpenDataDisclaimerModal={openDataDisclaimerModal}
+						onShowGridLoader={() => setGridLoader(true)}
 						handleMapStyleChange={onMapStyleChange}
 						searchValue={searchValue}
 						setSearchValue={setSearchValue}
@@ -1362,9 +751,10 @@ const DemoPage: React.FC = () => {
 				hideCancelButton
 			/>
 			<OpenDataConfirmationModal
-				open={show.openDataDisclaimerModal}
+				open={openDataDisclaimerModal}
 				onClose={() => {
-					setShow(s => ({ ...s, openDataDisclaimerModal: false, mapStyle: undefined }));
+					setOpenDataDisclaimerModal(false);
+					setTempMapStyle(undefined);
 					setMapProvider(currentMapProvider);
 				}}
 				onConfirm={() => setTimeout(() => handleOpenDataMapChange(), 0)}
@@ -1372,9 +762,10 @@ const DemoPage: React.FC = () => {
 				onConfirmationCheckboxOnChange={setDoNotAskOpenDataDisclaimer}
 			/>
 			<GrabConfirmationModal
-				open={show.grabDisclaimerModal}
+				open={grabDisclaimerModal}
 				onClose={() => {
-					setShow(s => ({ ...s, grabDisclaimerModal: false, mapStyle: undefined }));
+					setGrabDisclaimerModal(false);
+					setTempMapStyle(undefined);
 					setMapProvider(currentMapProvider);
 				}}
 				onConfirm={() => {
@@ -1402,7 +793,7 @@ const DemoPage: React.FC = () => {
 				onConfirm={() => {
 					setShow(s => ({ ...s, unauthSimulationExitModal: false, unauthGeofenceBox: false, unauthTrackerBox: false }));
 					setTimeout(() => {
-						handleGrabMapChange(show.mapStyle as GrabMapEnum);
+						handleGrabMapChange(tempMapStyle as GrabMapEnum);
 						window.location.reload();
 					}, 0);
 				}}
